backend/requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
Werkzeug==2.3.7
psutil==5.9.5
matplotlib==3.7.2
seaborn==0.12.2